# api-docs/swagger.yaml
openapi: 3.0.3
info:
  title: CCTV Smart Attendance System API
  description: AI-powered attendance management with facial recognition, PPE detection, salary/fine management, and HR analytics.
  version: 1.0.0
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api
    description: Development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from /auth/login
  parameters:
    pageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    limitParam:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
  schemas:
...
    PaginatedResponse:
      type: object
      properties:
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
...
    Salary:
      type: object
      properties:
        _id: { type: string }
        userId: { type: string }
        month: { type: integer, minimum: 1, maximum: 12 }
        year: { type: integer }
        baseSalary: { type: number }
        daysPresent: { type: number }
        daysAbsent: { type: number }
        workingDaysInMonth: { type: number }
        overtimeEarnings: { type: number }
        totalEarnings: { type: number }
        totalDeductions: { type: number }
        netSalary: { type: number }
        paymentStatus: { type: string, enum: [pending, processed, paid] }
...
    Shift:
      type: object
      properties:
        _id: { type: string }
        name: { type: string }
        startTime: { type: string, example: "09:00" }
        endTime: { type: string, example: "17:00" }
        gracePeriod: { type: integer }
        workingDays: { type: array, items: { type: integer, minimum: 0, maximum: 6 } }
...
security:
  - bearerAuth: []
paths:
...
  /salaries:
    get:
      tags: [Salaries]
      summary: List salary records
      parameters:
        - { $ref: '#/components/parameters/pageParam' }
        - { $ref: '#/components/parameters/limitParam' }
        - { name: month, in: query, schema: { type: integer, minimum: 1, maximum: 12 } }
        - { name: year, in: query, schema: { type: integer } }
      responses:
        200:
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/PaginatedResponse' }
                  - type: object
                    properties:
                      salaries: { type: array, items: { $ref: '#/components/schemas/Salary' } }
  /salaries/calculate:
    post:
      tags: [Salaries]
      summary: Calculate salary for a user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, month, year]
              properties:
                userId: { type: string }
                month: { type: integer, minimum: 1, maximum: 12 }
                year: { type: integer }
      responses:
        200:
          content:
            application/json:
              schema:
                type: object
                properties:
                  salary: { $ref: '#/components/schemas/Salary' }
  /salaries/process-all:
    post:
      tags: [Salaries]
      summary: Process salaries for all employees (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [month, year]
              properties:
                month: { type: integer, minimum: 1, maximum: 12 }
                year: { type: integer }
      responses:
        200:
          content:
            application/json:
              schema:
                type: object
                properties:
                  results: { type: array, items: { type: object } }
                  summary:
                    type: object
                    properties:
                      total: { type: integer }
                      succeeded: { type: integer }
                      failed: { type: integer }
  /salaries/{id}/payslip:
    get:
      tags: [Salaries]
      summary: Get payslip for a salary record
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        200:
          content:
            application/json:
              schema:
                type: object
                properties:
                  payslip: { type: object }

  # ─── Loans ─────────────────────────────────────────────────────
...
  /shifts:
    get:
      tags: [Shifts]
      summary: List shifts
      responses:
        200:
          content:
            application/json:
              schema:
                type: object
                properties:
                  shifts: { type: array, items: { $ref: '#/components/schemas/Shift' } }
    post:
      tags: [Shifts]
      summary: Create shift (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, startTime, endTime]
              properties:
                name: { type: string }
                startTime: { type: string, pattern: '^([01]\d|2[0-3]):[0-5]\d$' }
                endTime: { type: string, pattern: '^([01]\d|2[0-3]):[0-5]\d$' }
                gracePeriod: { type: integer, default: 15 }
                workingDays: { type: array, items: { type: integer, minimum: 0, maximum: 6 } }
      responses:
        201: { description: Shift created }
  /shifts/assign:
    post:
      tags: [Shifts]
      summary: Assign shift to user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, shiftId, effectiveFrom]
              properties:
                userId: { type: string }
                shiftId: { type: string }
                effectiveFrom: { type: string, format: date }
                effectiveTo: { type: string, format: date }
      responses:
        201: { description: Shift assigned }

  # ─── Analytics ─────────────────────────────────────────────────
...